rm server caching for dataset_metadata api endpoint [issue #166, pr #191]
In addition to setting cache control headers (to avoid caching of result), we also needed to not cache the data in the server's own memory cache. Note that this server caching logic is being removed entirely in the Cache Redesign epic, but this PR removes it in one place to specifically fix the Seamless collection/dataset dropdown problems.